Add dragbox util functions
Also fix some missing references (&) in ToolUtils.hpp

+++ include/gamelib/editor/tools/VertexTool.hpp
@@ -16,7 +16,7 @@ namespace gamelib
             void render(sf::RenderTarget& target);
 
         private:
-            int _selected;
+            size_t _selected;
             bool _snappoint;
     };
 }
